zng_color::gradient
Struct GradientRadius

pub struct GradientRadius {
    pub base: GradientRadiusBase,
    pub circle: bool,
    pub radii: Size,
}

The radial gradient radius length in both dimensions.

Fields
base: GradientRadiusBase

How the base length is calculated. The base length is the 100.pct() length.

circle: bool

If the gradient is circular or elliptical.

If true the radius is the same in both dimensions, if false the radius can be different.

radii: Size

The length of the rendered gradient stops.

Implementations

impl GradientRadius

pub fn closest_side(radius: impl Into<Size>) -> Self

Ellipse radii relative from center to the closest edge.

pub fn closest_corner(radius: impl Into<Size>) -> Self

Ellipse radii relative from center to the closest corner.

pub fn farthest_side(radius: impl Into<Size>) -> Self

Ellipse radii relative from center to the farthest edge.

pub fn farthest_corner(radius: impl Into<Size>) -> Self

Ellipse radii relative from center to the farthest corner.

pub fn circle(self) -> Self

Enable circular radius.

pub fn layout(&self, center: PxPoint) -> PxSize

Compute the radius in the current LAYOUT context.