zng_color

Struct Hsla

pub struct Hsla {
    pub hue: f32,
    pub saturation: f32,
    pub lightness: f32,
    pub alpha: f32,
}

HSL + alpha.

Equality

Equality is determined using about_eq with 0.001 granularity for hue and 0.00001 granularity for the others.

Fields
hue: f32

Hue color angle in the [0.0..=360.0] range.

saturation: f32

Saturation amount in the [0.0..=1.0] range, zero is gray, one is full color.

lightness: f32

Lightness amount in the [0.0..=1.0] range, zero is black, one is white.

alpha: f32

Alpha channel in the [0.0..=1.0] range, zero is invisible, one is opaque.

Implementations

impl Hsla

pub fn set_hue<H: Into<AngleDegree>>(&mut self, hue: H)

Sets the hue color angle.

The value is normalized to be in the [0.0..=360.0] range, that is 362.deg() becomes 2.0.

pub fn set_lightness<L: Into<Factor>>(&mut self, lightness: L)

Sets the lightness value.

pub fn set_saturation<S: Into<Factor>>(&mut self, saturation: S)

Sets the saturation value.

pub fn set_alpha<A: Into<Factor>>(&mut self, alpha: A)

Sets the alpha value.

pub fn with_hue<H: Into<AngleDegree>>(self, hue: H) -> Self

Returns a copy of this color with a new hue.

pub fn with_lightness<L: Into<Factor>>(self, lightness: L) -> Self

Returns a copy of this color with a new lightness.

pub fn with_saturation<S: Into<Factor>>(self, saturation: S) -> Self

Returns a copy of this color with a new saturation.

pub fn with_alpha<A: Into<Factor>>(self, alpha: A) -> Self

Returns a copy of this color with a new alpha.

pub fn slerp(self, to: Self, factor: Factor) -> Self

Interpolate in the LerpSpace::Hsla mode.

pub fn is_chromatic(self) -> bool

If the saturation is more than zero.

If false the color is achromatic, the hue value does not affect the color.

pub fn slerp_chromatic(self, to: Self, factor: Factor) -> Self

Interpolate in the LerpSpace::HslaChromatic mode.