Struct zng_color::filter::Filter

pub struct Filter { /* private fields */ }

A color filter or combination of filters.

Examples

use zng_color::filter::Filter;
use zng_layout::unit::*;

let filter = Filter::new_opacity(50.pct()).blur(3);

The example above creates a filter that lowers the opacity to 50% and blurs by 3px.

Implementations

impl Filter

pub fn opacity<A: Into<Factor>>(self, alpha: A) -> Self

Add an opacity adjustment to the filter, zero is fully transparent, one is the input transparency.

pub fn invert<A: Into<Factor>>(self, amount: A) -> Self

Add a color inversion filter, zero does not invert, one fully inverts.

pub fn blur<R: Into<Length>>(self, radius: R) -> Self

Add a blue effect to the filter, the blue radius is defined by a Length.

Relative lengths are calculated by the width of the available space.

pub fn sepia<A: Into<Factor>>(self, amount: A) -> Self

Add a sepia color effect to the filter, zero is the input color, one is the full desaturated brown look.

pub fn grayscale<A: Into<Factor>>(self, amount: A) -> Self

Add a grayscale color effect to the filter, zero is the input color, one if the full grayscale.

pub fn drop_shadow<O: Into<Point>, R: Into<Length>, C: Into<Rgba>>( self, offset: O, blur_radius: R, color: C ) -> Self

Add a drop-shadow to the effect.

pub fn brightness<A: Into<Factor>>(self, amount: A) -> Self

Add a brightness adjustment to the filter, zero removes all brightness, one is the input brightness.

pub fn contrast<A: Into<Factor>>(self, amount: A) -> Self

Add a contrast adjustment to the filter, zero removes all contrast, one is the input contrast.

pub fn saturate<A: Into<Factor>>(self, amount: A) -> Self

Add a saturation adjustment to the filter, zero fully desaturates, one is the input saturation.

pub fn hue_rotate<A: Into<AngleDegree>>(self, angle: A) -> Self

Add a filter that adds the angle to each color hue value.

pub fn flood<C: Into<Rgba>>(self, color: C) -> Self

Add a filter that fills the pixel space with color.

pub fn color_matrix<M: Into<ColorMatrix>>(self, matrix: M) -> Self

Custom filter.

The color matrix is in the format of SVG color matrix, [0..5] is the first matrix row.

impl Filter

pub const fn new() -> Filter

New default empty.

pub fn new_opacity<A: Into<Factor>>(alpha: A) -> Filter

New Filter::opacity.

pub fn new_invert<A: Into<Factor>>(amount: A) -> Filter

New Filter::invert.

pub fn new_blur<R: Into<Length>>(radius: R) -> Filter

New Filter::blur.

pub fn new_sepia<A: Into<Factor>>(amount: A) -> Filter

New Filter::sepia.

pub fn new_grayscale<A: Into<Factor>>(amount: A) -> Filter

New Filter::grayscale.

pub fn new_drop_shadow<O: Into<Point>, R: Into<Length>, C: Into<Rgba>>( offset: O, blur_radius: R, color: C ) -> Filter

New Filter::drop_shadow.

pub fn new_brightness<A: Into<Factor>>(amount: A) -> Filter

New Filter::brightness.

pub fn new_contrast<A: Into<Factor>>(amount: A) -> Filter

New Filter::contrast.

pub fn new_saturate<A: Into<Factor>>(amount: A) -> Filter

New Filter::saturate.

pub fn new_hue_rotate<A: Into<AngleDegree>>(angle: A) -> Filter

New Filter::hue_rotate.

pub fn new_flood<C: Into<Rgba>>(color: C) -> Filter

New Filter::flood.

pub fn new_color_matrix<M: Into<ColorMatrix>>(matrix: M) -> Filter

New Filter::color_matrix.

impl Filter

pub fn layout(&self) -> RenderFilter

Compute a RenderFilter in the current LAYOUT context.

Most filters convert one-to-one, effects that have a Length value use the layout context to calculate relative values.

Relative blur radius lengths are calculated using the constraints().fill_size().width value.

pub fn try_render(&self) -> Option<RenderFilter>

Compute a RenderFilter if the filter is not affected by layout.

pub fn needs_layout(&self) -> bool

Returns true if this filter is affected by the layout context where it is evaluated.