zng_color
Module gradient

Gradient types.

Macros

stops

Creates a GradientStops containing the arguments.

Structs

ColorStop

A color stop in a gradient.

GradientRadius

The radial gradient radius length in both dimensions.

GradientStops

Stops in a gradient.

Enums

ExtendMode

Specifies how to draw the gradient outside the first and last stop.

GradientRadiusBase

The radial gradient radius base length.

GradientStop

A stop in a gradient.

LinearGradientAxis

The angle or line that defines a linear gradient.

Type Aliases

RenderExtendMode

Gradient extend mode supported by the render.

RenderGradientStop

Computed GradientStop.