Trait zng_color::LightDarkVarExt

pub trait LightDarkVarExt {
    // Required methods
    fn rgba(&self) -> impl Var<Rgba>;
    fn rgba_map<T: VarValue>(
        &self,
        map: impl FnMut(Rgba) -> T + Send + 'static,
    ) -> impl Var<T>;
    fn rgba_into<T: VarValue + From<Rgba>>(&self) -> impl Var<T>;
    fn map_rgba(
        &self,
        map: impl FnMut(LightDark) -> LightDark + Send + 'static,
    ) -> impl Var<Rgba>;
    fn map_rgba_into<T: VarValue + From<Rgba>>(
        &self,
        map: impl FnMut(LightDark) -> LightDark + Send + 'static,
    ) -> impl Var<T>;
    fn shade_fct(&self, fct: impl Into<Factor>) -> impl Var<Rgba>;
    fn shade_fct_into<T: VarValue + From<Rgba>>(
        &self,
        fct: impl Into<Factor>,
    ) -> impl Var<T>;
    fn shade(&self, shade: i8) -> impl Var<Rgba>;
    fn shade_into<T: VarValue + From<Rgba>>(&self, shade: i8) -> impl Var<T>;
}

Extension methods for impl Var<LightDark>.

Required Methods

fn rgba(&self) -> impl Var<Rgba>

Gets a contextualized var that maps to LightDark::rgba.

fn rgba_map<T: VarValue>( &self, map: impl FnMut(Rgba) -> T + Send + 'static, ) -> impl Var<T>

Gets a contextualized var that maps to LightDark::rgba and map.

fn rgba_into<T: VarValue + From<Rgba>>(&self) -> impl Var<T>

Gets a contextualized var that maps to LightDark::rgba converted into T.

fn map_rgba( &self, map: impl FnMut(LightDark) -> LightDark + Send + 'static, ) -> impl Var<Rgba>

Gets a contextualized var that maps using map and then to LightDark::rgba.

fn map_rgba_into<T: VarValue + From<Rgba>>( &self, map: impl FnMut(LightDark) -> LightDark + Send + 'static, ) -> impl Var<T>

Gets a contextualized var that maps using map and then into T.

fn shade_fct(&self, fct: impl Into<Factor>) -> impl Var<Rgba>

Gets a contextualized var that maps to LightDark::shade_fct and then to LightDark::rgba.

fn shade_fct_into<T: VarValue + From<Rgba>>( &self, fct: impl Into<Factor>, ) -> impl Var<T>

Gets a contextualized var that maps to LightDark::shade_fct and then to LightDark::rgba and then into T.

fn shade(&self, shade: i8) -> impl Var<Rgba>

Gets a contextualized var that maps to LightDark::shade and then to LightDark::rgba.

+1 - Hovered.

+2 - Pressed.

fn shade_into<T: VarValue + From<Rgba>>(&self, shade: i8) -> impl Var<T>

Gets a contextualized var that maps to LightDark::shade and then to LightDark::rgba and then into T.

Object Safety

This trait is not object safe.